+++ CoreWebForms.sln
@@ -49,6 +49,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HtmlTextWriter", "src\HtmlT
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Compiler.Dynamic.Tests", "test\Compiler.Dynamic.Tests\Compiler.Dynamic.Tests.csproj", "{8392406E-4A1D-4312-9984-949F06F48C2A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Routing", "src\Routing\Routing.csproj", "{D010AD40-A04C-4E70-B7D2-3AD9712F5125}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -115,6 +117,10 @@ Global
 		{8392406E-4A1D-4312-9984-949F06F48C2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8392406E-4A1D-4312-9984-949F06F48C2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8392406E-4A1D-4312-9984-949F06F48C2A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D010AD40-A04C-4E70-B7D2-3AD9712F5125}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D010AD40-A04C-4E70-B7D2-3AD9712F5125}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D010AD40-A04C-4E70-B7D2-3AD9712F5125}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D010AD40-A04C-4E70-B7D2-3AD9712F5125}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -135,6 +141,7 @@ Global
 		{6BF8D296-2C69-4385-9D95-F28F32BFAF9D} = {D3D2945C-76BA-4DE3-90D2-259E2B1406C5}
 		{524EE349-51CA-4C4B-B397-97FCBDF7AB25} = {D3D2945C-76BA-4DE3-90D2-259E2B1406C5}
 		{8392406E-4A1D-4312-9984-949F06F48C2A} = {0E3C5885-15E6-4F5F-8D2C-088A0E255666}
+		{D010AD40-A04C-4E70-B7D2-3AD9712F5125} = {D3D2945C-76BA-4DE3-90D2-259E2B1406C5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B89024DA-2E51-49B1-8D52-651372AB5C83}
