--- src/Handlers/Adapters/HttpHandlerOptions.cs
@@ -1,8 +1,0 @@
-﻿// MIT License.
-
-namespace Microsoft.Extensions.DependencyInjection;
-
-public class HttpHandlerOptions
-{
-    public System.Web.Routing.RouteCollection Routes { get; } = new();
-}
