+++ src/Handlers/FeatureCollectionExtensions.cs
