+++ src/Handlers/HandlerServicesExtensions.cs
@@ -5,23 +5,17 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.SystemWebAdapters;
+using Microsoft.AspNetCore.SystemWebAdapters.HttpHandlers;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace Microsoft.Extensions.DependencyInjection;
 
 public static class HandlerServicesExtensions
 {
-    public static ISystemWebAdapterBuilder AddHttpHandlers(this ISystemWebAdapterBuilder services, Action<HttpHandlerOptions>? configure = null)
+    public static ISystemWebAdapterBuilder AddHttpHandlers(this ISystemWebAdapterBuilder services)
     {
         services.Services.TryAddSingleton<IHttpHandlerEndpointFactory, HttpHandlerEndpointFactory>();
         services.Services.AddTransient<IStartupFilter, HttpHandlerStartupFilter>();
-        services.Services.AddTransient(_ => new System.Web.Routing.RouteCollection());
-
-        if (configure is not null)
-        {
-            services.Services.AddOptions<HttpHandlerOptions>()
-                .Configure(configure);
-        }
 
         return services;
     }
