+++ src/Handlers/HttpHandlerEndpointFeature.cs
@@ -3,8 +3,9 @@
 using System.Web;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.SystemWebAdapters.Features;
 
-namespace Microsoft.AspNetCore.SystemWebAdapters;
+namespace Microsoft.AspNetCore.SystemWebAdapters.HttpHandlers;
 
 internal sealed class HttpHandlerEndpointFeature : IHttpHandlerFeature, IEndpointFeature
 {
