+++ src/Handlers/IHttpHandlerMetadata.cs
@@ -2,7 +2,7 @@
 
 using System.Web;
 
-namespace Microsoft.AspNetCore.SystemWebAdapters;
+namespace Microsoft.AspNetCore.SystemWebAdapters.HttpHandlers;
 
 public interface IHttpHandlerMetadata
 {
