+++ src/Handlers/SetHttpHandlerMiddleware.cs
@@ -2,8 +2,9 @@
 
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
+using Microsoft.AspNetCore.SystemWebAdapters.Features;
 
-namespace Microsoft.AspNetCore.SystemWebAdapters;
+namespace Microsoft.AspNetCore.SystemWebAdapters.HttpHandlers;
 
 internal sealed class SetHttpHandlerMiddleware
 {
