+++ src/Handlers/SystemWeb/HttpContextHandlerExtensions.cs
@@ -1,15 +1,12 @@
 // MIT License.
 
-using System.Runtime.CompilerServices;
-using System.Web.Routing;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.SystemWebAdapters;
-using Microsoft.Extensions.DependencyInjection;
+using Microsoft.AspNetCore.SystemWebAdapters.Features;
+using Microsoft.AspNetCore.SystemWebAdapters.HttpHandlers;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Primitives;
-
-using RouteData = System.Web.Routing.RouteData;
 
 namespace System.Web;
 
@@ -21,38 +18,11 @@ public static class HttpContextHandlerExtensions
     public static IHttpHandler? GetHandler(this HttpContext context)
         => ((HttpContextCore)context).Features.GetRequired<IHttpHandlerFeature>().Current;
 
-    public static RouteData GetRouteData(this HttpRequest request)
-    {
-        var coreRequest = ((HttpRequestCore)request);
-
-        if (coreRequest.HttpContext.Features.Get<RouteData>() is { } existing)
-        {
-            return existing;
-        }
-
-        var data = new RouteData(coreRequest.HttpContext.GetRouteData());
-
-        coreRequest.HttpContext.Features.Set<RouteData>(data);
-
-        return data;
-    }
-
-    public static RequestContext GetRequestContext(this HttpContextCore httpContext)
-    {
-        if (httpContext.Features.Get<RequestContext>() is { } existing)
-        {
-            return existing;
-        }
-
-        var routeData = GetRouteData(httpContext.Request);
-        var requestContext = new RequestContext(new HttpContextWrapper(httpContext), routeData);
-        httpContext.Features.Set<RequestContext>(requestContext);
-
-        return requestContext;
-    }
-
     public static IEndpointConventionBuilder MapHttpHandler<T>(this IEndpointRouteBuilder endpoints, string path)
        where T : IHttpHandler
+        => endpoints.MapHttpHandler(path, typeof(T));
+
+    public static IEndpointConventionBuilder MapHttpHandler(this IEndpointRouteBuilder endpoints, string path, Type type)
     {
         if (endpoints.DataSources.OfType<HttpHandlerManagerBuilder>().FirstOrDefault() is not { } existing)
         {
@@ -60,14 +30,14 @@ public static class HttpContextHandlerExtensions
             endpoints.DataSources.Add(existing);
         }
 
-        existing.Add<T>(path);
+        existing.Add(path, type);
 
         return existing;
     }
 
     private sealed class HttpHandlerManagerBuilder() : HttpHandlerEndpointConventionBuilder(new HttpHandlerManager())
     {
-        public void Add<T>(string path) => ((HttpHandlerManager)Manager).Add(HandlerEndpointBuilder.Create(path, typeof(T)));
+        public void Add(string path, Type type) => ((HttpHandlerManager)Manager).Add(HandlerEndpointBuilder.Create(path, type));
     }
 
     private sealed class HttpHandlerManager : IHttpHandlerManager
