+++ src/Handlers/SystemWeb/IRequiresSessionState.cs
