+++ src/WebForms/UI/Page.cs
@@ -2,6 +2,7 @@
 
 #nullable disable
 
+using WebForms.Routing;
 using System.Collections;
 using System.Collections.Specialized;
 using System.ComponentModel;
@@ -20,7 +21,6 @@ using System.Web.SessionState;
 using System.Web.UI.HtmlControls;
 using System.Web.UI.WebControls;
 using System.Web.Util;
-using Microsoft.Extensions.DependencyInjection;
 
 // Uncomment out this line to display rare field statistics at the end of the page
 //#define DISPLAYRAREFIELDSTATISTICS
