+++ src/WebForms/WebForms.csproj
@@ -37,6 +37,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Handlers\Handlers.csproj" />
     <ProjectReference Include="..\HtmlTextWriter\HtmlTextWriter.csproj" />
+    <ProjectReference Include="..\Routing\Routing.csproj" />
   </ItemGroup>
   
 </Project>
