+++ test/Webforms.Tests/PageTests.cs
@@ -3,19 +3,17 @@
 using System.Web.UI;
 using System.Web.UI.HtmlControls;
 using System.Web.UI.WebControls;
-using WebForms.Tests;
 using Xunit;
 
 namespace WebForms.Tests;
 
-[Collection(nameof(WebFormsBaseTest))]
-public class PageTests
+public class PageTests : HostedTestBase
 {
     [Fact]
     public async Task EmptyPage()
     {
         // Arrange/Act
-        var result = await TestUtil.RunPage<Page1>();
+        var result = await RunPage<Page1>();
 
         // Assert
         Assert.Empty(result);
@@ -25,7 +23,7 @@ public class PageTests
     public async Task CustomRender()
     {
         // Arrange/Act
-        var result = await TestUtil.RunPage<Page2>();
+        var result = await RunPage<Page2>();
 
         // Assert
         Assert.Equal("hello", result);
@@ -35,7 +33,7 @@ public class PageTests
     public async Task PageLoadAddControl()
     {
         // Arrange/Act
-        var result = await TestUtil.RunPage<Page3>();
+        var result = await RunPage<Page3>();
 
         // Assert
         Assert.Equal("hello", result);
@@ -45,7 +43,7 @@ public class PageTests
     public async Task PageWithForm()
     {
         // Arrange/Act
-        var result = await TestUtil.RunPage<Page4>();
+        var result = await RunPage<Page4>();
 
         // Assert
         Assert.Equal("<form method=\"post\" action=\"/path\"><div class=\"aspNetHidden\"</div></form>", result);
