--- test/Webforms.Tests/WebFormsBaseTest.cs
@@ -1,12 +1,0 @@
-// MIT License.
-
-using Xunit;
-
-namespace WebForms.Tests;
-
-[CollectionDefinition(nameof(WebFormsBaseTest),
-    DisableParallelization = true)]
-public class WebFormsBaseTest
-{
-
-}
